+++ mediawiki/includes/specials/SpecialMergeHistory.php
@@ -52,14 +52,14 @@ class SpecialMergeHistory extends SpecialPage {
 		$this->mTargetID = intval( $request->getVal( 'targetID' ) );
 		$this->mDestID = intval( $request->getVal( 'destID' ) );
 		$this->mTimestamp = $request->getVal( 'mergepoint' );
-		if( !preg_match( '/[0-9]{14}/', $this->mTimestamp ) ) {
+		if ( !preg_match( '/[0-9]{14}/', $this->mTimestamp ) ) {
 			$this->mTimestamp = '';
 		}
 		$this->mComment = $request->getText( 'wpComment' );
 
 		$this->mMerge = $request->wasPosted() && $this->getUser()->matchEditToken( $request->getVal( 'wpEditToken' ) );
 		// target page
-		if( $this->mSubmitted ) {
+		if ( $this->mSubmitted ) {
 			$this->mTargetObj = Title::newFromURL( $this->mTarget );
 			$this->mDestObj = Title::newFromURL( $this->mDest );
 		} else {
@@ -75,8 +75,8 @@ class SpecialMergeHistory extends SpecialPage {
 	 */
 	function preCacheMessages() {
 		// Precache various messages
-		if( !isset( $this->message ) ) {
-			$this->message['last'] = wfMsgExt( 'last', array( 'escape' ) );
+		if ( !isset( $this->message ) ) {
+			$this->message['last'] = $this->msg( 'last' )->escaped();
 		}
 	}
 
@@ -89,34 +89,37 @@ class SpecialMergeHistory extends SpecialPage {
 		$this->setHeaders();
 		$this->outputHeader();
 
-		if( $this->mTargetID && $this->mDestID && $this->mAction == 'submit' && $this->mMerge ) {
-			return $this->merge();
+		if ( $this->mTargetID && $this->mDestID && $this->mAction == 'submit' && $this->mMerge ) {
+			$this->merge();
+
+			return;
 		}
 
 		if ( !$this->mSubmitted ) {
 			$this->showMergeForm();
+
 			return;
 		}
 
 		$errors = array();
 		if ( !$this->mTargetObj instanceof Title ) {
-			$errors[] = wfMsgExt( 'mergehistory-invalid-source', array( 'parse' ) );
-		} elseif( !$this->mTargetObj->exists() ) {
-			$errors[] = wfMsgExt( 'mergehistory-no-source', array( 'parse' ),
+			$errors[] = $this->msg( 'mergehistory-invalid-source' )->parseAsBlock();
+		} elseif ( !$this->mTargetObj->exists() ) {
+			$errors[] = $this->msg( 'mergehistory-no-source', array( 'parse' ),
 				wfEscapeWikiText( $this->mTargetObj->getPrefixedText() )
-			);
+			)->parseAsBlock();
 		}
 
 		if ( !$this->mDestObj instanceof Title ) {
-			$errors[] = wfMsgExt( 'mergehistory-invalid-destination', array( 'parse' ) );
-		} elseif( !$this->mDestObj->exists() ) {
-			$errors[] = wfMsgExt( 'mergehistory-no-destination', array( 'parse' ),
+			$errors[] = $this->msg( 'mergehistory-invalid-destination' )->parseAsBlock();
+		} elseif ( !$this->mDestObj->exists() ) {
+			$errors[] = $this->msg( 'mergehistory-no-destination', array( 'parse' ),
 				wfEscapeWikiText( $this->mDestObj->getPrefixedText() )
-			);
+			)->parseAsBlock();
 		}
 
 		if ( $this->mTargetObj && $this->mDestObj && $this->mTargetObj->equals( $this->mDestObj ) ) {
-			$errors[] = wfMsgExt( 'mergehistory-same-destination', array( 'parse' ) );
+			$errors[] = $this->msg( 'mergehistory-same-destination' )->parseAsBlock();
 		}
 
 		if ( count( $errors ) ) {
@@ -125,7 +128,6 @@ class SpecialMergeHistory extends SpecialPage {
 		} else {
 			$this->showHistory();
 		}
-
 	}
 
 	function showMergeForm() {
@@ -137,25 +139,25 @@ class SpecialMergeHistory extends SpecialPage {
 			Xml::openElement( 'form', array(
 				'method' => 'get',
 				'action' => $wgScript ) ) .
-			'<fieldset>' .
-			Xml::element( 'legend', array(),
-				wfMsg( 'mergehistory-box' ) ) .
-			Html::hidden( 'title', $this->getTitle()->getPrefixedDbKey() ) .
-			Html::hidden( 'submitted', '1' ) .
-			Html::hidden( 'mergepoint', $this->mTimestamp ) .
-			Xml::openElement( 'table' ) .
-			'<tr>
-				<td>' . Xml::label( wfMsg( 'mergehistory-from' ), 'target' ) . '</td>
+				'<fieldset>' .
+				Xml::element( 'legend', array(),
+					$this->msg( 'mergehistory-box' )->text() ) .
+				Html::hidden( 'title', $this->getPageTitle()->getPrefixedDBkey() ) .
+				Html::hidden( 'submitted', '1' ) .
+				Html::hidden( 'mergepoint', $this->mTimestamp ) .
+				Xml::openElement( 'table' ) .
+				'<tr>
+				<td>' . Xml::label( $this->msg( 'mergehistory-from' )->text(), 'target' ) . '</td>
 				<td>' . Xml::input( 'target', 30, $this->mTarget, array( 'id' => 'target' ) ) . '</td>
 			</tr><tr>
-				<td>' . Xml::label( wfMsg( 'mergehistory-into' ), 'dest' ) . '</td>
+				<td>' . Xml::label( $this->msg( 'mergehistory-into' )->text(), 'dest' ) . '</td>
 				<td>' . Xml::input( 'dest', 30, $this->mDest, array( 'id' => 'dest' ) ) . '</td>
 			</tr><tr><td>' .
-			Xml::submitButton( wfMsg( 'mergehistory-go' ) ) .
-			'</td></tr>' .
-			Xml::closeElement( 'table' ) .
-			'</fieldset>' .
-			'</form>'
+				Xml::submitButton( $this->msg( 'mergehistory-go' )->text() ) .
+				'</td></tr>' .
+				Xml::closeElement( 'table' ) .
+				'</fieldset>' .
+				'</form>'
 		);
 	}
 
@@ -169,7 +171,7 @@ class SpecialMergeHistory extends SpecialPage {
 		$haveRevisions = $revisions && $revisions->getNumRows() > 0;
 
 		$out = $this->getOutput();
-		$titleObj = $this->getTitle();
+		$titleObj = $this->getPageTitle();
 		$action = $titleObj->getLocalURL( array( 'action' => 'submit' ) );
 		# Start the form here
 		$top = Xml::openElement(
@@ -182,40 +184,40 @@ class SpecialMergeHistory extends SpecialPage {
 		);
 		$out->addHTML( $top );
 
-		if( $haveRevisions ) {
+		if ( $haveRevisions ) {
 			# Format the user-visible controls (comment field, submission button)
 			# in a nice little table
 			$table =
 				Xml::openElement( 'fieldset' ) .
-				wfMsgExt( 'mergehistory-merge', array( 'parseinline' ),
-					$this->mTargetObj->getPrefixedText(), $this->mDestObj->getPrefixedText() ) .
-				Xml::openElement( 'table', array( 'id' => 'mw-mergehistory-table' ) ) .
+					$this->msg( 'mergehistory-merge', $this->mTargetObj->getPrefixedText(),
+						$this->mDestObj->getPrefixedText() )->parse() .
+					Xml::openElement( 'table', array( 'id' => 'mw-mergehistory-table' ) ) .
 					'<tr>
 						<td class="mw-label">' .
-							Xml::label( wfMsg( 'mergehistory-reason' ), 'wpComment' ) .
-						'</td>
-						<td class="mw-input">' .
-							Xml::input( 'wpComment', 50, $this->mComment, array( 'id' => 'wpComment' ) ) .
-						'</td>
+					Xml::label( $this->msg( 'mergehistory-reason' )->text(), 'wpComment' ) .
+					'</td>
+					<td class="mw-input">' .
+					Xml::input( 'wpComment', 50, $this->mComment, array( 'id' => 'wpComment' ) ) .
+					'</td>
 					</tr>
 					<tr>
 						<td>&#160;</td>
 						<td class="mw-submit">' .
-							Xml::submitButton( wfMsg( 'mergehistory-submit' ), array( 'name' => 'merge', 'id' => 'mw-merge-submit' ) ) .
-						'</td>
+					Xml::submitButton( $this->msg( 'mergehistory-submit' )->text(), array( 'name' => 'merge', 'id' => 'mw-merge-submit' ) ) .
+					'</td>
 					</tr>' .
-				Xml::closeElement( 'table' ) .
-				Xml::closeElement( 'fieldset' );
+					Xml::closeElement( 'table' ) .
+					Xml::closeElement( 'fieldset' );
 
 			$out->addHTML( $table );
 		}
 
 		$out->addHTML(
 			'<h2 id="mw-mergehistory">' .
-			wfMsgHtml( 'mergehistory-list' ) . "</h2>\n"
+				$this->msg( 'mergehistory-list' )->escaped() . "</h2>\n"
 		);
 
-		if( $haveRevisions ) {
+		if ( $haveRevisions ) {
 			$out->addHTML( $revisions->getNavigationBar() );
 			$out->addHTML( '<ul>' );
 			$out->addHTML( $revisions->getBody() );
@@ -225,8 +227,9 @@ class SpecialMergeHistory extends SpecialPage {
 			$out->addWikiMsg( 'mergehistory-empty' );
 		}
 
-		# Show relevant lines from the deletion log:
-		$out->addHTML( '<h2>' . htmlspecialchars( LogPage::logName( 'merge' ) ) . "</h2>\n" );
+		# Show relevant lines from the merge log:
+		$mergeLogPage = new LogPage( 'merge' );
+		$out->addHTML( '<h2>' . $mergeLogPage->getName()->escaped() . "</h2>\n" );
 		LogEventsList::showLogExtract( $out, 'merge', $this->mTargetObj );
 
 		# When we submit, go by page ID to avoid some nasty but unlikely collisions.
@@ -251,20 +254,22 @@ class SpecialMergeHistory extends SpecialPage {
 		$ts = wfTimestamp( TS_MW, $row->rev_timestamp );
 		$checkBox = Xml::radio( 'mergepoint', $ts, false );
 
+		$user = $this->getUser();
+
 		$pageLink = Linker::linkKnown(
 			$rev->getTitle(),
-			htmlspecialchars( $this->getLanguage()->timeanddate( $ts ) ),
+			htmlspecialchars( $this->getLanguage()->userTimeAndDate( $ts, $user ) ),
 			array(),
 			array( 'oldid' => $rev->getId() )
 		);
-		if( $rev->isDeleted( Revision::DELETED_TEXT ) ) {
+		if ( $rev->isDeleted( Revision::DELETED_TEXT ) ) {
 			$pageLink = '<span class="history-deleted">' . $pageLink . '</span>';
 		}
 
 		# Last link
-		if( !$rev->userCan( Revision::DELETED_TEXT, $this->getUser() ) ) {
+		if ( !$rev->userCan( Revision::DELETED_TEXT, $user ) ) {
 			$last = $this->message['last'];
-		} elseif( isset( $this->prevId[$row->rev_id] ) ) {
+		} elseif ( isset( $this->prevId[$row->rev_id] ) ) {
 			$last = Linker::linkKnown(
 				$rev->getTitle(),
 				$this->message['last'],
@@ -279,12 +284,13 @@ class SpecialMergeHistory extends SpecialPage {
 		$userLink = Linker::revUserTools( $rev );
 
 		$size = $row->rev_len;
-		if( !is_null( $size ) ) {
+		if ( !is_null( $size ) ) {
 			$stxt = Linker::formatRevisionSize( $size );
 		}
 		$comment = Linker::revComment( $rev );
 
-		return "<li>$checkBox ($last) $pageLink . . $userLink $stxt $comment</li>";
+		return Html::rawElement( 'li', array(),
+			$this->msg( 'mergehistory-revisionrow' )->rawParams( $checkBox, $last, $pageLink, $userLink, $stxt, $comment )->escaped() );
 	}
 
 	function merge() {
@@ -293,10 +299,10 @@ class SpecialMergeHistory extends SpecialPage {
 		# keep it consistent...
 		$targetTitle = Title::newFromID( $this->mTargetID );
 		$destTitle = Title::newFromID( $this->mDestID );
-		if( is_null( $targetTitle ) || is_null( $destTitle ) ) {
+		if ( is_null( $targetTitle ) || is_null( $destTitle ) ) {
 			return false; // validate these
 		}
-		if( $targetTitle->getArticleId() == $destTitle->getArticleId() ) {
+		if ( $targetTitle->getArticleID() == $destTitle->getArticleID() ) {
 			return false;
 		}
 		# Verify that this timestamp is valid
@@ -312,8 +318,9 @@ class SpecialMergeHistory extends SpecialPage {
 			__METHOD__
 		);
 		# Destination page must exist with revisions
-		if( !$maxtimestamp ) {
+		if ( !$maxtimestamp ) {
 			$this->getOutput()->addWikiMsg( 'mergehistory-fail' );
+
 			return false;
 		}
 		# Get the latest timestamp of the source
@@ -324,12 +331,13 @@ class SpecialMergeHistory extends SpecialPage {
 			__METHOD__
 		);
 		# $this->mTimestamp must be older than $maxtimestamp
-		if( $this->mTimestamp >= $maxtimestamp ) {
+		if ( $this->mTimestamp >= $maxtimestamp ) {
 			$this->getOutput()->addWikiMsg( 'mergehistory-fail' );
+
 			return false;
 		}
 		# Update the revisions
-		if( $this->mTimestamp ) {
+		if ( $this->mTimestamp ) {
 			$timewhere = "rev_timestamp <= {$this->mTimestamp}";
 			$timestampLimit = wfTimestamp( TS_MW, $this->mTimestamp );
 		} else {
@@ -349,68 +357,79 @@ class SpecialMergeHistory extends SpecialPage {
 		$haveRevisions = $dbw->selectField(
 			'revision',
 			'rev_timestamp',
-			array( 'rev_page' => $this->mTargetID  ),
+			array( 'rev_page' => $this->mTargetID ),
 			__METHOD__,
 			array( 'FOR UPDATE' )
 		);
-		if( !$haveRevisions ) {
-			if( $this->mComment ) {
-				$comment = wfMsgForContent(
+		if ( !$haveRevisions ) {
+			if ( $this->mComment ) {
+				$comment = $this->msg(
 					'mergehistory-comment',
 					$targetTitle->getPrefixedText(),
 					$destTitle->getPrefixedText(),
 					$this->mComment
-				);
+				)->inContentLanguage()->text();
 			} else {
-				$comment = wfMsgForContent(
+				$comment = $this->msg(
 					'mergehistory-autocomment',
 					$targetTitle->getPrefixedText(),
 					$destTitle->getPrefixedText()
-				);
+				)->inContentLanguage()->text();
 			}
-			$mwRedir = MagicWord::get( 'redirect' );
-			$redirectText = $mwRedir->getSynonym( 0 ) . ' [[' . $destTitle->getPrefixedText() . "]]\n";
-			$redirectPage = WikiPage::factory( $targetTitle );
-			$redirectRevision = new Revision( array(
-				'page'    => $this->mTargetID,
-				'comment' => $comment,
-				'text'    => $redirectText ) );
-			$redirectRevision->insertOn( $dbw );
-			$redirectPage->updateRevisionOn( $dbw, $redirectRevision );
 
-			# Now, we record the link from the redirect to the new title.
-			# It should have no other outgoing links...
-			$dbw->delete( 'pagelinks', array( 'pl_from' => $this->mDestID ), __METHOD__ );
-			$dbw->insert( 'pagelinks',
-				array(
-					'pl_from'      => $this->mDestID,
-					'pl_namespace' => $destTitle->getNamespace(),
-					'pl_title'     => $destTitle->getDBkey() ),
-				__METHOD__
-			);
+			$contentHandler = ContentHandler::getForTitle( $targetTitle );
+			$redirectContent = $contentHandler->makeRedirectContent( $destTitle );
+
+			if ( $redirectContent ) {
+				$redirectPage = WikiPage::factory( $targetTitle );
+				$redirectRevision = new Revision( array(
+					'title' => $targetTitle,
+					'page' => $this->mTargetID,
+					'comment' => $comment,
+					'content' => $redirectContent ) );
+				$redirectRevision->insertOn( $dbw );
+				$redirectPage->updateRevisionOn( $dbw, $redirectRevision );
+
+				# Now, we record the link from the redirect to the new title.
+				# It should have no other outgoing links...
+				$dbw->delete( 'pagelinks', array( 'pl_from' => $this->mDestID ), __METHOD__ );
+				$dbw->insert( 'pagelinks',
+					array(
+						'pl_from' => $this->mDestID,
+						'pl_namespace' => $destTitle->getNamespace(),
+						'pl_title' => $destTitle->getDBkey() ),
+					__METHOD__
+				);
+			} else {
+				// would be nice to show a warning if we couldn't create a redirect
+			}
 		} else {
 			$targetTitle->invalidateCache(); // update histories
 		}
 		$destTitle->invalidateCache(); // update histories
 		# Check if this did anything
-		if( !$count ) {
+		if ( !$count ) {
 			$this->getOutput()->addWikiMsg( 'mergehistory-fail' );
+
 			return false;
 		}
 		# Update our logs
 		$log = new LogPage( 'merge' );
 		$log->addEntry(
 			'merge', $targetTitle, $this->mComment,
-			array( $destTitle->getPrefixedText(), $timestampLimit )
+			array( $destTitle->getPrefixedText(), $timestampLimit ), $this->getUser()
 		);
 
-		$this->getOutput()->addHTML(
-			wfMsgExt( 'mergehistory-success', array('parseinline'),
-			$targetTitle->getPrefixedText(), $destTitle->getPrefixedText(), $count ) );
+		$this->getOutput()->addWikiMsg( 'mergehistory-success',
+			$targetTitle->getPrefixedText(), $destTitle->getPrefixedText(), $count );
 
 		wfRunHooks( 'ArticleMergeComplete', array( $targetTitle, $destTitle ) );
 
 		return true;
+	}
+
+	protected function getGroupName() {
+		return 'pagetools';
 	}
 }
 
@@ -447,9 +466,9 @@ class MergeHistoryPager extends ReverseChronologicalPager {
 			$batch->addObj( Title::makeTitleSafe( NS_USER_TALK, $row->user_name ) );
 
 			$rev_id = isset( $rev_id ) ? $rev_id : $row->rev_id;
-			if( $rev_id > $row->rev_id ) {
+			if ( $rev_id > $row->rev_id ) {
 				$this->mForm->prevId[$rev_id] = $row->rev_id;
-			} elseif( $rev_id < $row->rev_id ) {
+			} elseif ( $rev_id < $row->rev_id ) {
 				$this->mForm->prevId[$row->rev_id] = $rev_id;
 			}
 
@@ -460,6 +479,7 @@ class MergeHistoryPager extends ReverseChronologicalPager {
 		$this->mResult->seek( 0 );
 
 		wfProfileOut( __METHOD__ );
+
 		return '';
 	}
 
@@ -471,10 +491,11 @@ class MergeHistoryPager extends ReverseChronologicalPager {
 		$conds = $this->mConds;
 		$conds['rev_page'] = $this->articleID;
 		$conds[] = "rev_timestamp < {$this->maxTimestamp}";
+
 		return array(
 			'tables' => array( 'revision', 'page', 'user' ),
 			'fields' => array_merge( Revision::selectFields(), Revision::selectUserFields() ),
-			'conds'  => $conds,
+			'conds' => $conds,
 			'join_conds' => array(
 				'page' => Revision::pageJoinCond(),
 				'user' => Revision::userJoinCond() )
